aml_pipeline/concrete_ml/requirements.txt
fastapi==0.93.0
uvicorn==0.21.0
concrete-ml==0.6.0
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.0.2
joblib==1.2.0
onnx==1.13.0
xgboost==1.7.3
pydantic==1.8.2